# Ix: Improve IgnoreElements() performance