@akarnokd 2018-07-03
Added benchmarks for list and array sources.

#### Before

Method |       N |            Mean |          Error |         StdDev |  Gen 0 | Allocated |
------------ |--------:|----------------:|---------------:|---------------:|-------:|----------:|
Ignore |       1 |        47.36 ns |      0.7308 ns |      0.5706 ns | 0.0362 |     152 B |
IgnoreList |       1 |        45.95 ns |      0.3561 ns |      0.3157 ns | 0.0362 |     152 B |
IgnoreArray |       1 |        39.04 ns |      0.2439 ns |      0.2163 ns | 0.0343 |     144 B |
Ignore |      10 |        88.58 ns |      0.2004 ns |      0.1875 ns | 0.0361 |     152 B |
IgnoreList |      10 |       106.83 ns |      0.3054 ns |      0.2857 ns | 0.0362 |     152 B |
IgnoreArray |      10 |        76.90 ns |      0.3295 ns |      0.2921 ns | 0.0343 |     144 B |
Ignore |     100 |       515.36 ns |      1.3841 ns |      1.2947 ns | 0.0353 |     152 B |
IgnoreList |     100 |       624.86 ns |      2.6833 ns |      2.5099 ns | 0.0362 |     152 B |
IgnoreArray |     100 |       488.34 ns |      2.3555 ns |      2.2033 ns | 0.0343 |     144 B |
Ignore |    1000 |     4,727.26 ns |      7.3105 ns |      6.4806 ns | 0.0305 |     152 B |
IgnoreList |    1000 |     5,779.42 ns |     15.8669 ns |     14.8419 ns | 0.0305 |     152 B |
IgnoreArray |    1000 |     4,452.37 ns |     13.2589 ns |     12.4024 ns | 0.0305 |     144 B |
Ignore |   10000 |    46,737.52 ns |    112.0152 ns |    104.7791 ns |      - |     152 B |
IgnoreList |   10000 |    57,239.52 ns |    134.5450 ns |    119.2706 ns |      - |     152 B |
IgnoreArray |   10000 |    43,721.34 ns |     95.9446 ns |     85.0524 ns |      - |     144 B |
Ignore |  100000 |   466,920.08 ns |    760.3081 ns |    711.1926 ns |      - |     152 B |
IgnoreList |  100000 |   570,505.43 ns |  1,213.7155 ns |  1,135.3102 ns |      - |     152 B |
IgnoreArray |  100000 |   418,304.10 ns |  1,100.5126 ns |  1,029.4201 ns |      - |     144 B |
Ignore | 1000000 | 4,670,008.96 ns | 12,812.1047 ns | 11,984.4496 ns |      - |     152 B |
IgnoreList | 1000000 | 5,713,345.04 ns | 12,848.2853 ns | 12,018.2929 ns |      - |     152 B |
IgnoreArray | 1000000 | 4,155,854.58 ns | 15,964.1574 ns | 14,932.8813 ns |      - |     144 B |

#### After

Method |       N |            Mean |          Error |         StdDev |  Gen 0 | Allocated |
------------ |-------- |----------------:|---------------:|---------------:|-------:|----------:|
Ignore |       1 |        45.47 ns |      0.6528 ns |      0.5096 ns | 0.0362 |     152 B |
IgnoreList |       1 |        44.05 ns |      0.6448 ns |      0.6031 ns | 0.0362 |     152 B |
IgnoreArray |       1 |        37.14 ns |      0.1212 ns |      0.1012 ns | 0.0343 |     144 B |
Ignore |      10 |        70.78 ns |      0.1940 ns |      0.1720 ns | 0.0361 |     152 B |
IgnoreList |      10 |        75.68 ns |      0.3982 ns |      0.3725 ns | 0.0362 |     152 B |
IgnoreArray |      10 |        58.53 ns |      0.2932 ns |      0.2742 ns | 0.0343 |     144 B |
Ignore |     100 |       334.72 ns |      0.6147 ns |      0.5750 ns | 0.0358 |     152 B |
IgnoreList |     100 |       384.79 ns |      1.4730 ns |      1.3058 ns | 0.0362 |     152 B |
IgnoreArray |     100 |       277.80 ns |      1.9549 ns |      1.7330 ns | 0.0343 |     144 B |
Ignore |    1000 |     2,918.57 ns |     14.7415 ns |     12.3098 ns | 0.0343 |     152 B |
IgnoreList |    1000 |     3,435.35 ns |     19.7293 ns |     17.4895 ns | 0.0343 |     152 B |
IgnoreArray |    1000 |     2,397.52 ns |     14.3753 ns |     12.7434 ns | 0.0343 |     144 B |
Ignore |   10000 |    28,638.57 ns |     70.1554 ns |     62.1909 ns | 0.0305 |     152 B |
IgnoreList |   10000 |    33,869.47 ns |    169.7860 ns |    158.8179 ns |      - |     152 B |
IgnoreArray |   10000 |    23,449.30 ns |     46.1209 ns |     40.8850 ns | 0.0305 |     144 B |
Ignore |  100000 |   285,544.17 ns |    399.3706 ns |    373.5715 ns |      - |     152 B |
IgnoreList |  100000 |   337,894.56 ns |  2,040.0016 ns |  1,592.6989 ns |      - |     152 B |
IgnoreArray |  100000 |   234,222.61 ns |  1,485.8049 ns |  1,389.8227 ns |      - |     144 B |
Ignore | 1000000 | 2,855,214.10 ns | 10,413.2421 ns |  9,231.0670 ns |      - |     152 B |
IgnoreList | 1000000 | 3,383,301.22 ns | 17,448.2561 ns | 16,321.1080 ns |      - |     152 B |
IgnoreArray | 1000000 | 2,340,943.35 ns | 11,849.7648 ns | 11,084.2762 ns |      - |     144 B |