@onovotny 2018-07-03
Interesting how much worse list is than array.